feat: restructure application by removing unused CSS files and adding new components for layout and context management
